+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `className` option.
 - [Fix #118] Ignore `.textcomplete('destory')` on non-initialized elements.
+- [Fix #119] Trigger completer with the current text by default.
 
 ## `v0.3.2` - Sep 16, 2014